build.py: Support building for multiple SYSTEM_PROFILE & APPLICATION_PROFILE pairs
The new '-T' or '--target-pair' argument can be used to specify one or
more system profile & application profile pairs to build, using the
format 'SYSTEM_PROFILE:APPLICATION_PROFILE'. For example, to build a
guest image and a host image for raspberrypi3:

    ./scripts/build.py -M raspberrypi3 -T guest:minimal -T native:host

The old style of '-S' and '-A' arguments are still supported for
backwards compatibility. If both types of argument are given, the target
pair defined by the '-S' and '-A' arguments is built first.

Signed-off-by: Paul Barker <pbarker@oryx-linux.org>