README.md: Update README.md
The README was updated to remove non-existent links ahead of the
release, and changed the main documentation link such as links to
CODE_OF_CONDUCT.md and CONTRIBUTING.md. Once those documents are
added to mion, the README will be updated.

Additionally, markdownlint was run against README, and minor changes
were made, such as removal of trailing whitespace and blanks around
headers was added.

This addresses mion-docs #32

Signed-off-by: Katrina Prosise <igorina@toganlabs.com>