onie-only-build.sh: Add a simple build script.
Because of how we need to pass VENDOR into bitbake end users
need to either pass in BB_ENV_EXTRAWHITE (danger) or use mc_build.sh

This is a simplier version of mc_build.sh and should NOT be seen as
the canonical way to build mion but is simply there as a convenience.
Anything else? Use mc_build and multiconfig.

Signed-off-by: Eilís Ní Fhlannagáin <pidge@toganlabs.com>