Cronie Improvements
BitBake is capable of building multiple packages in one invocation, by providing
more than one package name. Cronie currently quotes the package argument, so
supplying multiple packages will appear as a single package to bitbake. This
commit reverts the quotes.

It also adds the following arguments:

   -f ) Force operation
   -c ) Command to apply

This is useful if you want to clean the shared state or otherwise clean a bunch
of packages, without having to run "cronie.sh -e ... " and grabbing all of the
environment setup.

Signed-off-by: G Davey <gdavey@committhis.co.uk>